refactor(binding-kafka): rename remaining api Response classes with Kafka prefix
CreateTopicsResponse(V7FW) and DeleteTopicsResponse(V6FW) become
KafkaCreateTopicsResponse(V7FW) and KafkaDeleteTopicsResponse(V6FW), matching
the Kafka* naming already used by every other class in the binding-kafka api
package. Updates the two consumers, KafkaClientCreateTopicsFactory and
McpKafkaProxyFactory, accordingly.
