+++ runtime/binding-kafka/src/test/java/io/aklivity/zilla/runtime/binding/kafka/api/KafkaDeleteTopicsResponseV6FWTest.java
@@ -21,15 +21,13 @@ import static org.junit.Assert.assertTrue;
 
 import org.junit.Test;
 
-import io.aklivity.zilla.runtime.binding.kafka.api.CreateTopicsResponse.Config;
-import io.aklivity.zilla.runtime.binding.kafka.api.CreateTopicsResponse.Kind;
-import io.aklivity.zilla.runtime.binding.kafka.api.CreateTopicsResponse.Topic;
+import io.aklivity.zilla.runtime.binding.kafka.api.KafkaDeleteTopicsResponse.Topic;
 import io.aklivity.zilla.runtime.common.agrona.buffer.DirectBufferEx;
 import io.aklivity.zilla.runtime.common.agrona.buffer.UnsafeBufferEx;
 
-public class CreateTopicsResponseV7FWTest
+public class KafkaDeleteTopicsResponseV6FWTest
 {
-    // body bytes only, as verified against the real KafkaClientCreateTopicsFactory v7 wire decoder input
+    // body bytes only, as verified against the real KafkaClientDeleteTopicsFactory v6 wire decoder input
     // (the response header's correlationId is decoded separately and excluded here)
     private static final byte[] BODY = new byte[]
     {
@@ -40,23 +38,11 @@ public class CreateTopicsResponseV7FWTest
         0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
         0x00, 0x00,
         0x00,
-        0x00, 0x00, 0x00, 0x01,
-        0x00, 0x01,
-        0x02,
-        0x0f, 'c', 'l', 'e', 'a', 'n', 'u', 'p', '.', 'p', 'o', 'l', 'i', 'c', 'y',
-        0x07, 'd', 'e', 'l', 'e', 't', 'e',
-        0x00,
-        0x01,
         0x00,
         0x00,
-        0x00,
-        0x0a, 's', 'n', 'a', 'p', 's', 'h', 'o', 't', 's',
         0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00,
-        0x00,
-        0x00, 0x00, 0x00, 0x01,
-        0x00, 0x01,
-        0x01,
+        0x00, 0x03,
+        0x10, 't', 'o', 'p', 'i', 'c', ' ', 'n', 'o', 't', ' ', 'f', 'o', 'u', 'n', 'd',
         0x00,
         0x00
     };
@@ -70,9 +56,9 @@ public class CreateTopicsResponseV7FWTest
     }
 
     @Test
-    public void shouldDecodeCreateTopicsV7Response()
+    public void shouldDecodeDeleteTopicsV6Response()
     {
-        CreateTopicsResponseV7FW response = new CreateTopicsResponseV7FW();
+        KafkaDeleteTopicsResponseV6FW response = new KafkaDeleteTopicsResponseV6FW();
 
         DirectBufferEx buffer = new UnsafeBufferEx(BODY);
         response.wrap(buffer, 0, buffer.capacity());
@@ -81,36 +67,20 @@ public class CreateTopicsResponseV7FWTest
         assertEquals(2, response.topicCount());
 
         assertTrue(response.hasNext());
-        assertEquals(Kind.TOPIC, response.next());
-
-        Topic events = response.topic();
+        Topic events = response.next();
         assertEquals("events", asString(events.buffer(), events.nameOffset(), events.nameLength()));
         assertEquals(0L, events.topicIdMostSigBits());
         assertEquals(0L, events.topicIdLeastSigBits());
         assertEquals(0, events.error());
         assertEquals(-1, events.messageLength());
-        assertEquals(1, events.numPartitions());
-        assertEquals(1, events.replicationFactor());
-        assertEquals(1, events.configCount());
 
         assertTrue(response.hasNext());
-        assertEquals(Kind.CONFIG, response.next());
-
-        Config cleanupPolicy = response.config();
-        assertEquals("cleanup.policy", asString(cleanupPolicy.buffer(), cleanupPolicy.nameOffset(), cleanupPolicy.nameLength()));
-        assertEquals("delete", asString(cleanupPolicy.buffer(), cleanupPolicy.valueOffset(), cleanupPolicy.valueLength()));
-        assertFalse(cleanupPolicy.readOnly());
-        assertEquals(1, cleanupPolicy.configSource());
-        assertFalse(cleanupPolicy.isSensitive());
-
-        assertTrue(response.hasNext());
-        assertEquals(Kind.TOPIC, response.next());
-
-        Topic snapshots = response.topic();
-        assertEquals("snapshots", asString(snapshots.buffer(), snapshots.nameOffset(), snapshots.nameLength()));
-        assertEquals(1, snapshots.numPartitions());
-        assertEquals(1, snapshots.replicationFactor());
-        assertEquals(0, snapshots.configCount());
+        Topic missing = response.next();
+        assertEquals(-1, missing.nameLength());
+        assertEquals(0L, missing.topicIdMostSigBits());
+        assertEquals(0L, missing.topicIdLeastSigBits());
+        assertEquals(3, missing.error());
+        assertEquals("topic not found", asString(missing.buffer(), missing.messageOffset(), missing.messageLength()));
 
         assertFalse(response.hasNext());
         assertEquals(BODY.length, response.limit());
